requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
PyMySQL==1.1.0
email-validator==2.1.1
cryptography==42.0.5
python-dotenv==1.0.1
